refactor: iwantmyname
